+++ eras/shelley/impl/cardano-ledger-shelley.cabal
@@ -125,3 +125,20 @@ library
 
     if flag(asserts)
         ghc-options: -fno-ignore-asserts
+
+library testlib
+    exposed-modules:  Test.Cardano.Ledger.Shelley.Arbitrary
+    visibility:       public
+    hs-source-dirs:   testlib
+    default-language: Haskell2010
+    ghc-options:
+        -Wall -Wcompat -Wincomplete-record-updates
+        -Wincomplete-uni-patterns -Wredundant-constraints -Wpartial-fields
+        -Wunused-packages
+
+    build-depends:
+        base >=4.12 && <4.17,
+        cardano-ledger-core:{cardano-ledger-core, testlib},
+        cardano-ledger-shelley,
+        generic-random,
+        mtl
