+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Serialisation/Generators.hs
@@ -15,18 +15,11 @@ module Test.Cardano.Ledger.Shelley.Serialisation.Generators () where
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.API (ShelleyTxBody (ShelleyTxBody))
 import Cardano.Ledger.Shelley.LedgerState (PPUPPredFailure)
-import Cardano.Ledger.Shelley.PParams
 import qualified Cardano.Ledger.Shelley.Rules as STS
-import Control.Monad.Identity (Identity)
-import Data.Maybe.Strict (StrictMaybe)
 import Generic.Random (genericArbitraryU)
 import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (Mock)
 import Test.Cardano.Ledger.Shelley.Serialisation.EraIndepGenerators ()
-import Test.QuickCheck (
-  Arbitrary,
-  arbitrary,
-  shrink,
- )
+import Test.QuickCheck (Arbitrary (..))
 
 {-------------------------------------------------------------------------------
   ShelleyEra Generators
@@ -36,7 +29,7 @@ import Test.QuickCheck (
 -------------------------------------------------------------------------------}
 
 instance
-  (EraTxOut era, Mock (EraCrypto era), Arbitrary (TxOut era)) =>
+  (EraTxOut era, Arbitrary (PParamsUpdate era), Arbitrary (TxOut era)) =>
   Arbitrary (ShelleyTxBody era)
   where
   arbitrary =
@@ -61,13 +54,3 @@ instance
   where
   arbitrary = genericArbitraryU
   shrink _ = []
-
-instance Era era => Arbitrary (ShelleyPParams Identity era) where
-  arbitrary = genericArbitraryU
-
-instance Era era => Arbitrary (ShelleyPParams StrictMaybe era) where
-  arbitrary = genericArbitraryU
-
-deriving instance (Era era, Arbitrary (PParamsHKD Identity era)) => Arbitrary (PParams era)
-
-deriving instance (Era era, Arbitrary (PParamsHKD StrictMaybe era)) => Arbitrary (PParamsUpdate era)
